refactor: 방향키 이벤트 훅 분리
